pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scentmap"
version = "0.1.0"
description = "Scentmap service"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.109.0",
  "uvicorn[standard]==0.27.0",
  "pydantic==2.5.3",
  "pydantic-settings==2.1.0",
  "python-dotenv==1.0.0",
  "typing-extensions==4.9.0",
  "psycopg2-binary==2.9.9",
  "httpx==0.26.0",
  "python-multipart==0.0.6",
  "orjson==3.9.15",
]

[tool.setuptools]
packages = ["scentmap", "scentmap.app", "scentmap.app.api", "scentmap.app.schemas", "scentmap.app.services"]
